refactor(KlaviyoSwift): relocate state-mirror + event-publish bridge off KlaviyoInternal (MAGE-832) (#636)
* refactor(KlaviyoSwift): relocate state-store mirror off KlaviyoInternal (MAGE-832)


* refactor(KlaviyoSwift): relocate event publish off KlaviyoInternal (MAGE-832)


* style: tidy comment + swiftlint for state-mirror relocation (MAGE-832)


* test(KlaviyoSwift): hoist @MainActor to class level in relocated tests (MAGE-832)

Addresses review feedback on #636: annotate EventPublishTests and
SharedStoreMirrorTests with class-level @MainActor instead of per-method.


---------
