+++ Sources/KlaviyoSwift/KlaviyoInternal.swift
@@ -29,8 +29,6 @@ package enum KlaviyoInternal {
     private static var apiKeyCancellable: Cancellable?
     private static let apiKeySubject = CurrentValueSubject<APIKeyResult, Never>(.failure(.notInitialized))
 
-    private static var sharedStoresCancellable: Cancellable?
-
     // MARK: - API Key methods
 
     // Setup the profile data subject to receive updates from the state publisher
@@ -98,12 +96,6 @@ package enum KlaviyoInternal {
 
     // Setup the profile data subject to receive updates from the state publisher
     private static func setupProfileDataSubject() {
-        // Defensive, idempotent attach of the shared-store mirror. The mirror is normally
-        // established at `initialize(with:)` and left alive for the SDK's lifetime — it is NOT
-        // torn down on In-App Forms teardown (see `resetProfileDataSubject()`), so this is a
-        // safety net rather than a required re-attach hook.
-        setupSharedStores()
-
         // Only set up the subscription if it hasn't already been set up
         guard profileDataCancellable == nil else { return }
 
@@ -117,24 +109,6 @@ package enum KlaviyoInternal {
             }
             .removeDuplicates()
             .subscribe(profileDataSubject)
-    }
-
-    /// Mirrors initialized SDK state into the shared `KlaviyoCore` stores so other modules
-    /// (Forms, Location) can observe identity and API key without importing `KlaviyoSwift`.
-    package static func setupSharedStores() {
-        guard sharedStoresCancellable == nil else { return }
-        sharedStoresCancellable = klaviyoSwiftEnvironment.statePublisher()
-            .filter { $0.initalizationState == .initialized }
-            .map { (identity: $0.identity, apiKey: $0.apiKey) }
-            .removeDuplicates(by: { $0.identity == $1.identity && $0.apiKey == $1.apiKey })
-            // TCA dispatches state changes on the main thread, so these two sequential
-            // store writes are observed together rather than torn. Write the config
-            // before the identity: IdentityStore.update(_:) notifies observers
-            // synchronously, so an identity observer must not see a stale apiKey.
-            .sink { identity, apiKey in
-                SDKConfigStore.shared.update(KlaviyoConfig(apiKey: apiKey))
-                IdentityStore.shared.update(identity)
-            }
     }
 
     /// Fetches the current profile data once.
@@ -172,31 +146,15 @@ package enum KlaviyoInternal {
         profileDataCancellable = nil
         profileDataSubject.send(.failure(.notInitialized))
         // NOTE: deliberately does NOT tear down the shared-store mirror or clear the stores.
-        // The mirror is global SDK state (established once by `setupSharedStores()` at
+        // The mirror is global SDK state (established once by `SharedStoreMirror.setup()` at
         // initialize) that KlaviyoForms/KlaviyoLocation observe directly. Clearing it on
         // In-App Forms teardown would leave the stores empty after an unregister → re-register
         // cycle even while the SDK stays initialized, since consumers now read the stores
-        // directly and no longer re-trigger `setupSharedStores()`. (Broader teardown
+        // directly and no longer re-trigger `SharedStoreMirror.setup()`. (Broader teardown
         // decoupling is tracked in MAGE-834.)
     }
 
-    /// Tears down the shared-store mirror and clears the stores. **Test-only isolation helper** —
-    /// production never tears the mirror down (see `resetProfileDataSubject()`); this exists so
-    /// tests that call `setupSharedStores()` start from a detached, empty state.
-    package static func resetSharedStores() {
-        sharedStoresCancellable?.cancel()
-        sharedStoresCancellable = nil
-        SDKConfigStore.shared.update(KlaviyoConfig())
-        IdentityStore.shared.update(ProfileData())
-    }
-
     // MARK: - Profile Event methods
-
-    /// Publishes an event to the KlaviyoCore event bus after enriching it with metadata.
-    /// - Parameter event: the profile event to publish
-    internal static func publishEvent(_ event: Event) {
-        EventBus.shared.publish(enrichEventWithMetadata(event))
-    }
 
     // TODO(MAGE-834): fold reset semantics into the KlaviyoCore stores/bus.
     /// No-op. The event-reset mechanism moves onto KlaviyoCore in MAGE-834.
@@ -205,21 +163,6 @@ package enum KlaviyoInternal {
     /// Clears the event bus replay buffer to ensure clean state between tests.
     package static func clearEventBuffer() {
         EventBus.shared.clearBuffer()
-    }
-
-    /// Enriches an event with metadata (device info, SDK info, etc.)
-    /// - Parameter event: The event to enrich
-    /// - Returns: A new Event with metadata appended to properties
-    private static func enrichEventWithMetadata(_ event: Event) -> Event {
-        let enrichedProperties = event.properties.appendMetadataToProperties() ?? event.properties
-        return Event(
-            name: event.metric.name,
-            properties: enrichedProperties,
-            identifiers: event.identifiers,
-            value: event.value,
-            time: event.time,
-            uniqueId: event.uniqueId
-        )
     }
 
     // MARK: - Aggregate Events methods
